+++ src/test/java/io/sentry/unit/autoinstall/graphql/Graphql22AutoInstallTest.kt
@@ -76,7 +76,7 @@ class Graphql22AutoInstallTest {
         assertTrue {
             fixture.logger.capturedMessage ==
                 "sentry-graphql-22 won't be installed because the current " +
-                "version is lower than the minimum supported version 22.0.0"
+                "version (21.9.0) is lower than the minimum supported version 22.0.0"
         }
 
         assertTrue(fixture.dependencies.none { it.artifactId == "sentry-graphql-22" })