+++ src/test/java/io/sentry/unit/autoinstall/spring/Spring7AutoInstallTest.kt
@@ -1,7 +1,7 @@
 package io.sentry.unit.autoinstall.spring
 
 import io.sentry.autoinstall.AutoInstallState
-import io.sentry.autoinstall.spring.Spring5InstallStrategy
+import io.sentry.autoinstall.spring.Spring7InstallStrategy
 import io.sentry.unit.fakes.CapturingTestLogger
 import org.apache.maven.model.Dependency
 import org.eclipse.aether.artifact.Artifact
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test
 import org.slf4j.Logger
 import kotlin.test.assertTrue
 
-class Spring5AutoInstallTest {
+class Spring7AutoInstallTest {
     class Fixture {
         val logger = CapturingTestLogger()
         val dependencies = ArrayList<Dependency>()
@@ -19,9 +19,9 @@ class Spring5AutoInstallTest {
 
         fun getSut(
             installSpring: Boolean = true,
-            springVersion: String = "5.1.2",
-        ): Spring5InstallStrategy {
-            installState = AutoInstallState("6.25.2")
+            springVersion: String = "7.0.0",
+        ): Spring7InstallStrategy {
+            installState = AutoInstallState("8.21.0")
             installState.isInstallSpring = installSpring
 
             resolvedArtifacts.add(
@@ -33,70 +33,67 @@ class Spring5AutoInstallTest {
                 ),
             )
 
-            return Spring5InstallStrategyImpl(logger)
+            return Spring7InstallStrategyImpl(logger)
         }
     }
 
     private val fixture = Fixture()
 
     @Test
-    fun `when sentry-spring is a direct dependency logs a message and does nothing`() {
+    fun `when sentry-spring-7 is a direct dependency logs a message and does nothing`() {
         val sut = fixture.getSut(installSpring = false)
-
         sut.install(fixture.dependencies, fixture.resolvedArtifacts, fixture.installState)
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-spring won't be installed because it was already " +
+                "sentry-spring-7 won't be installed because it was already " +
                 "installed directly"
         }
 
-        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring" })
+        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring-7" })
     }
 
     @Test
     fun `when spring version is too low logs a message and does nothing`() {
-        val sut = fixture.getSut(springVersion = "5.1.1")
-
+        val sut = fixture.getSut(springVersion = "6.7.4")
         sut.install(fixture.dependencies, fixture.resolvedArtifacts, fixture.installState)
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-spring won't be installed because the current " +
-                "version (5.1.1) is lower than the minimum supported version 5.1.2"
+                "sentry-spring-7 won't be installed because the current " +
+                "version (6.7.4) is lower than the minimum supported version 7.0.0-M1"
         }
 
-        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring" })
+        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring-7" })
     }
 
     @Test
     fun `when spring version is too high logs a message and does nothing`() {
-        val sut = fixture.getSut(springVersion = "6.0.0")
+        val sut = fixture.getSut(springVersion = "8.0.0")
 
         sut.install(fixture.dependencies, fixture.resolvedArtifacts, fixture.installState)
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-spring won't be installed because the current " +
-                "version (6.0.0) is higher than the maximum supported version 5.9999.9999"
+                "sentry-spring-7 won't be installed because the current " +
+                "version (8.0.0) is higher than the maximum supported version 7.9999.9999"
         }
 
-        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring" })
+        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-spring-7" })
     }
 
     @Test
-    fun `installs sentry-spring with info message`() {
+    fun `installs sentry-spring-7 with info message`() {
         val sut = fixture.getSut()
-
         sut.install(fixture.dependencies, fixture.resolvedArtifacts, fixture.installState)
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-spring was successfully installed with version: 6.25.2"
+                "sentry-spring-7 was successfully installed with version: 8.21.0"
         }
 
-        assertTrue(fixture.dependencies.any { it.artifactId == "sentry-spring" })
+        assertTrue(fixture.dependencies.any { it.artifactId == "sentry-spring-7" })
     }
 
-    private class Spring5InstallStrategyImpl(logger: Logger) : Spring5InstallStrategy(logger)
+    private class Spring7InstallStrategyImpl(logger: Logger) : Spring7InstallStrategy(logger)
 }